feat(mla): add oss-fuzz support - Configure mla-oss-fuzz with libFuzzer target - Update FUZZING.md documentation - Factorize fuzzing code in mla-fuzz crate
