+++ .github/workflows/test.yml
@@ -77,7 +77,7 @@ jobs:
         # RUSTC_BOOTSTRAP=1 allows use of -Z build-std on stable toolchain
         # Required because win7 targets aren't officially supported in stable
         RUSTC_BOOTSTRAP: "${{ contains(matrix.target, 'win7') && '1' || '' }}"
-      run: cargo build --release --all --exclude mla-fuzz-afl --verbose ${{ matrix.cargo_arg }}
+      run: cargo build --release --all --exclude mla-fuzz-afl --exclude mla-oss-fuzz --verbose ${{ matrix.cargo_arg }}
     - name: Run tests (Windows)
       # as rustdoc.exe has stack overflows with current tests (likely due to ml-dsa expansion) in mla/lib.rs
       if: matrix.os == 'windows-2025-vs2026'
@@ -85,10 +85,10 @@ jobs:
         # RUSTC_BOOTSTRAP=1 allows use of -Z build-std on stable toolchain
         # Required because win7 targets aren't officially supported in stable
         RUSTC_BOOTSTRAP: "${{ contains(matrix.target, 'win7') && '1' || '' }}"     
-      run: cargo test --all --lib --bins --tests --exclude mla-fuzz-afl --release --verbose ${{ matrix.cargo_arg }}
+      run: cargo test --all --lib --bins --tests --exclude mla-fuzz-afl --exclude mla-oss-fuzz --release --verbose ${{ matrix.cargo_arg }}
     - name: Run tests (Linux/macOS)
       if: matrix.os != 'windows-2025-vs2026'    
-      run: cargo test --all --exclude mla-fuzz-afl --release --verbose
+      run: cargo test --all --exclude mla-fuzz-afl --exclude mla-oss-fuzz --release --verbose
 
   long-tests:
     runs-on: ubuntu-26.04
@@ -106,7 +106,7 @@ jobs:
     steps:
     - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
     - name: Run long tests
-      run: cargo test --all-features --all --exclude mla-fuzz-afl --exclude mla-bindings-c --release
+      run: cargo test --all-features --all --exclude mla-fuzz-afl --exclude mla-bindings-c --exclude mla-oss-fuzz --release
 
   test-bindings-c-cpp-linux:
     runs-on: ubuntu-26.04
@@ -166,7 +166,7 @@ jobs:
       - name: Test AFL build
         env:
           RUSTFLAGS: -D warnings
-        run: cargo afl build
+        run: cargo afl build --workspace --exclude mla-oss-fuzz
 
   fmt:
     runs-on: ubuntu-26.04
